requirements.txt
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
et-xmlfile==1.0.1
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
kiwisolver==1.3.1
line-bot-sdk==1.17.0
lxml==4.6.1
MarkupSafe==1.1.1
matplotlib==3.3.2
numpy==1.18.4
oauthlib==3.1.0
openpyxl==3.0.5
pandas==1.0.3
pandas-datareader==0.9.0
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0.1
TA-Lib==0.4.19
urllib3==1.25.9
Werkzeug==1.0.1
xlrd==1.2.0
